Repair dividends: fine-tune logic + unit tests
Fix detecting div_adj_exceeds_prices.
Fine-tune fixing div_too_big.
Unit tests.
